+++ core/amber/src/main/scala/edu/uci/ics/texera/workflow/common/workflow/WorkflowCacheRewriter.scala
@@ -1,10 +1,13 @@
 package edu.uci.ics.texera.workflow.common.workflow
 
-import edu.uci.ics.texera.web.service.WorkflowCacheChecker
+import edu.uci.ics.texera.Utils.objectMapper
+import edu.uci.ics.texera.web.service.{ExecutionsMetadataPersistService, WorkflowCacheChecker}
 import edu.uci.ics.texera.workflow.common.storage.OpResultStorage
 import edu.uci.ics.texera.workflow.common.workflow.LogicalPlan.schemaPropagationCheck
 import edu.uci.ics.texera.workflow.operators.sink.SinkOpDesc
+import edu.uci.ics.texera.workflow.operators.sink.managed.ProgressiveSinkOpDesc
 import edu.uci.ics.texera.workflow.operators.source.cache.CacheSourceOpDesc
+import edu.uci.ics.texera.workflow.operators.visualization.VisualizationConstants
 
 import scala.collection.mutable
 import scala.collection.mutable.ArrayBuffer
@@ -15,8 +18,7 @@ object WorkflowCacheRewriter {
       logicalPlan: LogicalPlan,
       lastCompletedPlan: Option[LogicalPlan],
       storage: OpResultStorage,
-      opsToReuseCache: Set[String],
-      opsUsedCache: mutable.Set[String]
+      opsToReuseCache: Set[String]
   ): LogicalPlan = {
     val validCachesFromLastExecution =
       new WorkflowCacheChecker(lastCompletedPlan, logicalPlan).getValidCacheReuse
@@ -73,10 +75,58 @@ object WorkflowCacheRewriter {
       resultPlan.terminalOperators.forall(o => resultPlan.getOperator(o).isInstanceOf[SinkOpDesc])
     )
 
-    // output the operators that have used cache to outside
-    opsCanUseCache.foreach(opId => opsUsedCache.add(opId))
+
+    // assign sink storage to the logical plan after cache rewrite
+    // as it will be converted to the actual physical plan
+    assignSinkStorage(logicalPlan, storage, opsCanUseCache)
+    // also assign sink storage to the original logical plan, as the original logical plan
+    // will be used to be compared to the subsequent runs
+    assignSinkStorage(resultPlan, storage, opsCanUseCache)
     resultPlan
 
+  }
+
+  private def assignSinkStorage(
+                                 logicalPlan: LogicalPlan,
+                                 storage: OpResultStorage,
+                                 reuseStorageSet: Set[String] = Set()
+                               ): Unit = {
+    // create a JSON object that holds pointers to the workflow's results in Mongo
+    // TODO in the future, will extract this logic from here when we need pointers to the stats storage
+    val resultsJSON = objectMapper.createObjectNode()
+    val sinksPointers = objectMapper.createArrayNode()
+    // assign storage to texera-managed sinks before generating exec config
+    logicalPlan.operators.foreach {
+      case o@(sink: ProgressiveSinkOpDesc) =>
+        val storageKey = sink.getUpstreamId.getOrElse(o.operatorID)
+        // due to the size limit of single document in mongoDB (16MB)
+        // for sinks visualizing HTMLs which could possibly be large in size, we always use the memory storage.
+        val storageType = {
+          if (sink.getChartType.contains(VisualizationConstants.HTML_VIZ)) OpResultStorage.MEMORY
+          else OpResultStorage.defaultStorageMode
+        }
+        if (reuseStorageSet.contains(storageKey) && storage.contains(storageKey)) {
+          sink.setStorage(storage.get(storageKey))
+        } else {
+          sink.setStorage(
+            storage.create(
+              o.context.executionID + "_",
+              storageKey,
+              storageType
+            )
+          )
+          sink.getStorage.setSchema(logicalPlan.outputSchemaMap(o.operatorIdentifier).head)
+          // add the sink collection name to the JSON array of sinks
+          sinksPointers.add(o.context.executionID + "_" + storageKey)
+        }
+      case _ =>
+    }
+    // update execution entry in MySQL to have pointers to the mongo collections
+    resultsJSON.set("results", sinksPointers)
+    ExecutionsMetadataPersistService.updateExistingExecutionVolumePointers(
+      logicalPlan.context.executionID,
+      resultsJSON.toString
+    )
   }
 
 }
