+++ core/amber/src/main/scala/edu/uci/ics/texera/workflow/common/workflow/WorkflowCompiler.scala
@@ -3,17 +3,12 @@ package edu.uci.ics.texera.workflow.common.workflow
 import edu.uci.ics.amber.engine.architecture.controller.Workflow
 import edu.uci.ics.amber.engine.architecture.scheduling.WorkflowPipelinedRegionsBuilder
 import edu.uci.ics.amber.engine.common.virtualidentity.WorkflowIdentity
-import edu.uci.ics.texera.Utils.objectMapper
 import edu.uci.ics.texera.web.model.websocket.request.LogicalPlanPojo
-import edu.uci.ics.texera.web.service.ExecutionsMetadataPersistService
 import edu.uci.ics.texera.web.storage.JobStateStore
 import edu.uci.ics.texera.web.workflowruntimestate.FatalErrorType.COMPILATION_ERROR
 import edu.uci.ics.texera.workflow.common.WorkflowContext
 import edu.uci.ics.texera.workflow.common.storage.OpResultStorage
-import edu.uci.ics.texera.workflow.operators.sink.managed.ProgressiveSinkOpDesc
-import edu.uci.ics.texera.workflow.operators.visualization.VisualizationConstants
 
-import scala.collection.mutable
 import scala.collection.mutable.ArrayBuffer
 
 class WorkflowCompiler(
@@ -21,75 +16,6 @@ class WorkflowCompiler(
     workflowContext: WorkflowContext,
     jobStateStore: JobStateStore
 ) {
-  def compileLogicalPlan(
-      opResultStorage: OpResultStorage,
-      lastCompletedJob: Option[LogicalPlan] = Option.empty
-  ): (LogicalPlan, LogicalPlan) = {
-    var logicalPlan: LogicalPlan = LogicalPlan(logicalPlanPojo, workflowContext)
-
-    logicalPlan = SinkInjectionTransformer.transform(logicalPlanPojo.opsToViewResult, logicalPlan)
-
-    val opsUsedCache = new mutable.HashSet[String]()
-    val rewrittenLogicalPlan =
-      WorkflowCacheRewriter.transform(
-        logicalPlan,
-        lastCompletedJob,
-        opResultStorage,
-        logicalPlanPojo.opsToReuseResult.toSet,
-        opsUsedCache
-      )
-
-    // assign sink storage to the logical plan after cache rewrite
-    // as it will be converted to the actual physical plan
-    assignSinkStorage(rewrittenLogicalPlan, opResultStorage, opsUsedCache.toSet)
-    // also assign sink storage to the original logical plan, as the original logical plan
-    // will be used to be compared to the subsequent runs
-    assignSinkStorage(logicalPlan, opResultStorage, opsUsedCache.toSet)
-    (logicalPlan, rewrittenLogicalPlan)
-  }
-
-  private def assignSinkStorage(
-      logicalPlan: LogicalPlan,
-      storage: OpResultStorage,
-      reuseStorageSet: Set[String] = Set()
-  ): Unit = {
-    // create a JSON object that holds pointers to the workflow's results in Mongo
-    // TODO in the future, will extract this logic from here when we need pointers to the stats storage
-    val resultsJSON = objectMapper.createObjectNode()
-    val sinksPointers = objectMapper.createArrayNode()
-    // assign storage to texera-managed sinks before generating exec config
-    logicalPlan.operators.foreach {
-      case o @ (sink: ProgressiveSinkOpDesc) =>
-        val storageKey = sink.getUpstreamId.getOrElse(o.operatorID)
-        // due to the size limit of single document in mongoDB (16MB)
-        // for sinks visualizing HTMLs which could possibly be large in size, we always use the memory storage.
-        val storageType = {
-          if (sink.getChartType.contains(VisualizationConstants.HTML_VIZ)) OpResultStorage.MEMORY
-          else OpResultStorage.defaultStorageMode
-        }
-        if (reuseStorageSet.contains(storageKey) && storage.contains(storageKey)) {
-          sink.setStorage(storage.get(storageKey))
-        } else {
-          sink.setStorage(
-            storage.create(
-              o.context.executionID + "_",
-              storageKey,
-              storageType
-            )
-          )
-          sink.getStorage.setSchema(logicalPlan.outputSchemaMap(o.operatorIdentifier).head)
-          // add the sink collection name to the JSON array of sinks
-          sinksPointers.add(o.context.executionID + "_" + storageKey)
-        }
-      case _ =>
-    }
-    // update execution entry in MySQL to have pointers to the mongo collections
-    resultsJSON.set("results", sinksPointers)
-    ExecutionsMetadataPersistService.updateExistingExecutionVolumePointers(
-      logicalPlan.context.executionID,
-      resultsJSON.toString
-    )
-  }
 
   def compile(
       workflowId: WorkflowIdentity,
@@ -105,30 +31,50 @@ class WorkflowCompiler(
       )
     }
 
-    val (logicalPlan0, logicalPlan) = compileLogicalPlan(opResultStorage, lastCompletedJob)
+    // generate an original LogicalPlan. The logical plan is the injected with all necessary sinks
+    //  this plan will be compared in subsequent runs to check which operator can be replaced
+    //  by cache.
+    val originalLogicalPlan: LogicalPlan = LogicalPlan(logicalPlanPojo, workflowContext)
 
-    val physicalPlan = PhysicalPlan(logicalPlan)
+    // the cache-rewritten LogicalPlan. It is considered to be equivalent with the original plan.
+    val rewrittenLogicalPlan = WorkflowCacheRewriter.transform(
+        originalLogicalPlan,
+        lastCompletedJob,
+        opResultStorage,
+        logicalPlanPojo.opsToReuseResult.toSet
+      )
 
-    // create pipelined regions.
+
+    // the PhysicalPlan with topology expanded.
+    val physicalPlan = PhysicalPlan(rewrittenLogicalPlan)
+
+    // generate an ExecutionPlan with regions.
+    //  currently, WorkflowPipelinedRegionsBuilder is the only ExecutionPlan generator.
     val executionPlan = new WorkflowPipelinedRegionsBuilder(
       workflowId,
-      logicalPlan,
+      rewrittenLogicalPlan,
       physicalPlan,
-      new MaterializationRewriter(logicalPlan.context, opResultStorage)
+      new MaterializationRewriter(rewrittenLogicalPlan.context, opResultStorage)
     ).buildPipelinedRegions()
 
-    // assign link strategies
-    val partitioningPlan = new PartitionEnforcer(physicalPlan, executionPlan).enforcePartition()
+    // TODO: add resource allocator to incorporate PartitioningPlan below
+
+    // generate a PartitioningPlan, describing partitions between physical operators
+    val partitioningPlan = new PartitionEnforcer(physicalPlan).enforcePartition()
 
     // assert all source layers to have 0 input ports
     physicalPlan.getSourceOperators.foreach { sourceLayer =>
       assert(physicalPlan.getLayer(sourceLayer).inputPorts.isEmpty)
     }
+    // assert all sink layers to have 0 output ports
+    physicalPlan.getSinkOperators.foreach{
+      sinkLayer => assert(physicalPlan.getLayer(sinkLayer).outputPorts.isEmpty)
+    }
 
     new Workflow(
       workflowId,
-      logicalPlan0,
-      logicalPlan,
+      originalLogicalPlan,
+      rewrittenLogicalPlan,
       physicalPlan,
       executionPlan,
       partitioningPlan
